ctf/crypto/ProteinCookies2/challenge/requirements.txt
Flask
Pycryptodome